Fix S057 function in alias facts
